tts-controller/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
httpx==0.25.2
pyyaml==6.0.1
requests==2.31.0
docker==6.1.3

# 测试依赖
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
